requirements.txt
# OSFormer Core Requirements (Optimized for NVIDIA A100 - Fixed Versions)

# Core Deep Learning Framework
# Recommended for A100: CUDA 12.1 stable
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1

# Computer Vision & Image Processing
opencv-python==4.8.1.78
pillow==10.1.0

# Scientific Computing & Data Analysis
numpy==1.24.3
pandas==2.1.3
scipy==1.11.4

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
tqdm==4.66.1
pyyaml==6.0.1
psutil==5.9.6
